% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm_script.R
\name{my_lm}
\alias{my_lm}
\title{Linear Regression}
\usage{
my_lm(response, covariates, alpha, method)
}
\arguments{
\item{response}{A \code{vector} response variable y of your data set of interest.}

\item{covariates}{A \code{matrix} containing the explanatory variable x of your data set of interest.}

\item{alpha}{A \code{numeric} set to determine an estimate of confidence intervals. The most common confidence interval is 95 percent, with a level of significance, alpha = 0.05. Generally, alpha is between 0.01 and 0.1.}

\item{method}{A \code{character} used to determine the fit; method = "asymptotic" or method = "bootstrap".}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{beta}{Estimated value of beta hat}
     \item{sigma2}{Estimate of the residual variance}
     \item{variance_beta}{Estimate of the variance of the estimated beta}
     \item{ci}{Estimate of the confidence interval based on alpha}
     \item{mspe}{Estimate of how well the model predicts the response variable}
     \item{ssm}{Model sum of squares}
     \item{sse}{Error sum of squares, quantify how much the data points vary around the estimated regression line, y.hat}
     \item{f.stat}{Calculated f statistic for use in f test and model fitting}
     \item{residual}{??}
     \item{y.hat}{Estimated slope of the regression line??}
}
}
\description{
This is a package for performing linear regressions and also associated statistical outputs. Confidence intervals and mean square prediction error can be computed. F-tests are able to be conducted and publication quality plots of data are able to be generated. 
}
\examples{

my_lm(hubble$y, hubble$x, 0.05, "bootstrap")

my_lm(hubble$y, hubble$x, 0.05, "asymptotic")
}
